--- a/.babelrc
+++ b/.babelrc
@@ -9,6 +9,7 @@
   ],
   "plugins": [
     "transform-react-inline-elements",
-    "transform-react-constant-elements"
+    "transform-react-constant-elements",
+    "syntax-dynamic-import"
   ]
 }