--- a/package.json
+++ b/package.json
@@ -134,7 +134,9 @@
     "asar": "~0.13.0",
     "babel": "^6.1.18",
     "babel-core": "^6.3.15",
+    "babel-eslint": "^8.0.3",
     "babel-loader": "^7.1.1",
+    "babel-plugin-syntax-dynamic-import": "^6.18.0",
     "babel-plugin-transform-react-constant-elements": "^6.4.0",
     "babel-plugin-transform-react-inline-elements": "^6.4.0",
     "babel-polyfill": "^6.3.14",
@@ -205,6 +207,7 @@
       "muon",
       "postMessage"
     ],
+    "parser": "babel-eslint",
     "ignore": [
       "app/extensions/**",
       "app/browser/ads/adDivCandidates.js",